// api/src/daily-task/types/daily-task.types.ts
export enum TaskScope {
  PERSONAL_GENERAL = 'personal_general',
  PERSONAL_HERO = 'personal_hero',
}

export enum TaskMetric {
  KILLS = 'kills',
  ASSISTS = 'assists',
  LAST_HITS = 'last_hits',
  TOWER_KILLS = 'tower_kills',
  HERO_DAMAGE = 'hero_damage',
  HEALING = 'healing',
  TOTAL_GOLD_EARNED = 'total_gold_earned',
  DAMAGE_TAKEN = 'damage_taken',
  STUN_DURATION = 'stun_duration',
  ROSHAN_KILLS = 'roshan_kills',
}